+++ tools/build_linux.py
@@ -10,7 +10,7 @@ import tarfile
 from pathlib import Path
 
 APP_VERSION = "1.6.0"
-RC_VERSION = "rc.3"
+RC_VERSION = "rc.4"
 APP_NAME = "dicodePing"